requirements.txt
# Core dependencies
playwright>=1.40.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-playwright>=0.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Utilities
colorama>=0.4.6  # For colored terminal output
click>=8.1.0     # For CLI interface